requirements.txt
matplotlib>=3.0.0
numpy>=1.20.0
pandas>=1.3.0
plotly>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
streamlit>=1.0.0
sqlalchemy>=1.4.0